Refs #4, fisrt approach to pypi compatible app
