+++ setup.cfg
@@ -1,0 +1,8 @@
+[metadata]
+# This includes the license file(s) in the wheel.
+license_files = LICENSE.txt
+
+[bdist_wheel]
+# 1 to build for several python versions
+# 0 to build to just the supported one
+universal=0
